+++ src/Exceptions/FileNotFoundException.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace AhmadMayahi\GoogleVision\Exceptions;
+
+use Exception;
+
+class FileNotFoundException extends Exception
+{
+}